backend/pyproject.toml
[project]
name = "research-pipeline"
version = "0.1.0"
description = "AI-powered research pipeline with LangGraph agents"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",          # async PostgreSQL driver used by SQLAlchemy
    # LangChain + LangGraph
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",  # DuckDuckGoSearchResults lives here
    "langgraph>=0.3.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    # Pydantic (used by FastAPI and LangChain)
    "pydantic>=2.0.0",
    # Web search
    "duckduckgo-search>=7.0.0",
    "ddgs>=9.14.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ✅ Tells hatchling your code lives in the app/ folder
[tool.hatch.build.targets.wheel]
packages = ["app"]

# ✅ Modern syntax replacing the deprecated tool.uv.dev-dependencies
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
